Refactor: split SettingsModal into domain panels (#76) (#106)
* refactor(settings): extract shared PanelHeading primitive

Pull the small h3 heading used by every Settings panel into its own
file so the soon-to-be-extracted domain panels can import it without
each duplicating the markup. No behavioural change — same DOM, same
classes.


* refactor(settings): extract GeneralPanel into its own file

Move General settings (folder sort, hidden folders, trash mode, share
defaults, "Show welcome tab") out of SettingsModal.tsx. The parent
shell will switch over to the import once every panel has been
extracted; for now the inline copy still owns rendering.


* refactor(settings): extract AppearancePanel into its own file

Move theme preset/token pickers and the Fonts section (FontsSection +
FontSlotRow + normalizeForPicker helper) out of SettingsModal.tsx.
No behavioural change.


* refactor(settings): extract EditorPanel into its own file

Move the editor settings (preview mode, autocorrect, reopen tabs,
task density, lenient-done-today) out of SettingsModal.tsx.


* refactor(settings): extract SidebarPanel into its own file

Move hidden-sidebar-tabs management (list + Show buttons) out of
SettingsModal.tsx.


* refactor(settings): extract LocalFolderPanel into its own file

Move File System Access folder pickup + InFolderGitSection (init,
remote, commit, push) + the local ExclamationTriangleIconUnsupported
helper out of SettingsModal.tsx. No behavioural change.


* refactor(settings): extract GitHubPanel into its own file

Move GitHub-sync settings out of SettingsModal.tsx, along with the
inline fields it owns: VaultGitignoreField, GitignoreOverlayField,
VaultEncryptionField, ResetToRemoteField. The dynamic imports for
vault-key / reset-to-remote / gitignore-sync stay intact.


* refactor(settings): extract BetaPanel and AboutPanel into their own files

Both are short and have no shared state with the other panels, so they
land together. Beta surfaces feature flags; About has the version line,
help / GitHub / issue links, the report-bug button, and the launch
EmailSignup.


* refactor(settings): add panels barrel index for the new domain files

Single import surface — the shell will pull every panel from
'./settings/panels' instead of declaring eight separate paths.


* refactor(settings): drop inline panels in SettingsModal, wire to extracted files (#76)

Replace the eight inline panel definitions (General, Appearance, Editor,
Sidebar, LocalFolder, GitHub, Beta, About) plus their helpers
(ShowWelcomeButton, FontsSection, FontSlotRow, normalizeForPicker,
InFolderGitSection, VaultEncryptionField, ResetToRemoteField,
VaultGitignoreField, GitignoreOverlayField, ExclamationTriangleIconUnsupported,
PanelHeading) with imports from ./settings/panels.

What stays in the shell: modal lifecycle, the category nav rail,
the search-across-settings input + result list (#54), the
SettingsFooterBar, and the CategoryPanel switcher. File shrinks from
1984 lines to ~330 — small enough to read end-to-end.

Behaviour preserved: same DOM, same data-testids, same Zustand
selectors. settingsModal.test.tsx + settingsSearch.test.tsx pass
unchanged.


---------
